Server/requirements.txt
# Core backend
Flask
flask-cors
python-dotenv

# LangChain & components
langchain
langchain-community


# Embeddings
sentence-transformers
transformers


# Vector store
chromadb

# Optional: URL document loading support
unstructured
requests
